+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyComponent.java
@@ -26,8 +26,8 @@ import org.odk.collect.android.configure.qr.ShowQRCodeFragment;
 import org.odk.collect.android.formentry.ODKView;
 import org.odk.collect.android.formentry.QuitFormDialogFragment;
 import org.odk.collect.android.formentry.saving.SaveFormProgressDialogFragment;
-import org.odk.collect.android.formmanagement.matchexactly.SyncFormsTaskSpec;
-import org.odk.collect.android.formmanagement.previouslydownloaded.AutoUpdateTaskSpec;
+import org.odk.collect.android.backgroundwork.SyncFormsTaskSpec;
+import org.odk.collect.android.backgroundwork.AutoUpdateTaskSpec;
 import org.odk.collect.android.fragments.BarCodeScannerFragment;
 import org.odk.collect.android.fragments.DataManagerList;
 import org.odk.collect.android.fragments.MapBoxInitializationFragment;
